New commit

+++ src/controllers/cartController.js
@@ -3,6 +3,22 @@ const Cart = require("../models/cartModel");
 const User = require("../models/userModel");
 const Product = require("../models/productModel");
 
+const checkQuantity = (products) => {
+  let quantity = 0;
+  for (let index = 0; index < products.length; index++) {
+    const product = products[index];
+    quantity += product.quantity;
+  }
+  return quantity;
+};
+const getSubtotal = (products) => {
+  let subtotal = 0;
+  for (let index = 0; index < products.length; index++) {
+    const product = products[index];
+    subtotal += product.subtotal;
+  }
+};
+
 const addToCart = async (req, res) => {
   try {
     const userID = req.params.id;
@@ -30,18 +46,21 @@ const addToCart = async (req, res) => {
       console.log("user has a cart");
       // check if product already exists in cart
       let product = cart.products.find(
-        (item) => item.productId === req.body.productId // transverse the array to find if product to be added in the array already exist in the cart
+        (item) => item.product._id === req.body.productId // transverse the array to find if product to be added in the array already exist in the cart
       );
       // ADD TO CART SECTION
       // CHECK IF PRODUCT ALREADY IN CART
       if (product) {
         console.log("product already exists in cart");
 
-        let quantity = parseInt(req.body.quantity); // convert product quantity to a number
+        let quantity = parseInt(req.body.quantity);
+        // convert product quantity to a number
 
-        product.quantity += quantity || 1; // it's add specified quantity of product or adds 1 if quantity not specified
+        product.quantity += quantity || 1;
+        // it's add specified quantity of product or adds 1 if quantity not specified
         product.subtotal = product.quantity * product.price;
-
+        cart.itemCount = checkQuantity(cart.products);
+        cart.totalPrice = getSubtotal(cart.products);
         cart.save(cart, (err, cart) => {
           if (err) {
             return res.json({
@@ -61,12 +80,12 @@ const addToCart = async (req, res) => {
         console.log(" user has cart but doesnt have product in his cart");
 
         cart.products.push({
-          productId: req.body.productId,
-          quantity: req.body.quantity,
-          name: req.body.name,
-          price: req.body.price,
-          subtotal: req.body.price * req.body.quantity,
+          product: req.body.productId,
+          quantity: req.body.quantity || 1,
+          subtotal: productExist.price * (req.body.quantity || 1),
         });
+        cart.itemCount = checkQuantity(cart.products);
+        cart.quantity = getSubtotal(cart.products);
         cart.save((err, cart) => {
           if (err) {
             return res.json({
@@ -92,13 +111,13 @@ const addToCart = async (req, res) => {
         products: [
           {
             productId: req.body.productId,
-            quantity: req.body.quantity,
-            name: req.body.name,
-            price: req.body.price,
-            subtotal: req.body.price * req.body.quantity,
+            quantity: req.body.quantity || 1,
+            price: productExist.price,
+            subtotal: productExist.price * (req.body.quantity || 1),
           },
         ],
-        active: true,
+        itemCount: req.body.quantity || 1,
+        totalPrice: productExist.price * (req.body.quantity || 1),
       });
       cart.save((err, cart) => {
         if (err) {
@@ -134,7 +153,7 @@ const removeFromCart = async (req, res) => {
 
     const productExist = await Product.findOne({
       productId: req.body.productId,
-    }); //check if product to be added to the cart exists in the store
+    }); //check if product to be removed from the cart exists in the store
 
     if (!productExist) {
       return res
@@ -149,7 +168,7 @@ const removeFromCart = async (req, res) => {
 
       // check if product already exists in cart
       let product = cart.products.find(
-        (item) => item.productId === req.body.productId // transverse the array to find if product to be added in the array already exist in the cart
+        (item) => item.product._id === req.body.productId // transverse the array to find if product to be added in the array already exist in the cart
       );
       // REMOVE TO CART SECTION
       // CHECK IF PRODUCT ALREADY IN CART
@@ -159,16 +178,14 @@ const removeFromCart = async (req, res) => {
         let quantity = parseInt(req.body.quantity); // convert product quantity to a number
         if (cart.products.length == 1 && product.quantity == 1) {
           await Cart.findOneAndDelete({ userId: user._id });
-          return res.status(200).json({
-            success: true,
-            message: "product deleted from cart successfully",
-          });
         }
         if (product.quantity == 1) {
           cart.products.splice(cart.products.indexOf(req.body.productId), 1);
         } else {
           product.quantity -= quantity || 1; // it's subtracts specified quantity of product or subtracts 1 if quantity not specified
           product.subtotal = product.quantity * product.price;
+          cart.itemCount = checkQuantity(cart.products);
+          cart.totalPrice = getSubtotal(cart.products);
         }
         cart.save(cart, (err, cart) => {
           if (err) {
@@ -179,7 +196,7 @@ const removeFromCart = async (req, res) => {
           } else {
             return res.json({
               success: true,
-              message: "Cart updated",
+              message: "product deleted from cart successfully",
               cart: cart,
             });
           }
@@ -193,9 +210,9 @@ const removeFromCart = async (req, res) => {
 
     // if user has no cart, create one
     else {
-      res
+      return res
         .status(409)
-        .json({ success: false, message: "cannot delete from empty cart  " });
+        .json({ success: false, message: "this user has no cart" });
     }
   } catch (error) {
     console.log(error);
@@ -229,8 +246,20 @@ const getCart = async (req, res) => {
 const deleteCart = async (req, res) => {
   try {
     // check if cart exists
-    await Cart.findOneAndDelete;
-    ({ userId: req.params });
+    await Cart.findOneAndDelete({ userId: req.params }).then((cart, err) => {
+      if (err) {
+        res.status(400).json({
+          success: false,
+          message: "error occurred while deleting this user's cart",
+        });
+      } else {
+        return res.status(200).json({
+          success: true,
+          message: "cart deleted successfully",
+          cart: cart,
+        });
+      }
+    });
   } catch (error) {
     console.log(error);
   }
